+++ src/scr/libs/StuntCarRacerSharpLib/Screens/GameOverScreen.cs
@@ -1,4 +1,4 @@
-// 'Stunt Car Racer - The Sharp Kind' - Andy Hawkins 2026.
+﻿// 'Stunt Car Racer - The Sharp Kind' - Andy Hawkins 2026.
 // 'Stunt Car Racer Remake' - sourceforge.net/projects/stuntcarremake.
 // Stunt Car Racer (C) Geoff Crammond / MicroStyle / MicroProse 1989.
 
@@ -37,7 +37,7 @@ internal sealed class GameOverScreen : IGameScreen
 
     public void Draw()
     {
-        _race.DrawWorld(showOpponent: true);
+        _race.DrawWorld(showOpponent: true, showPlayer: true);
         _race.DrawHud(gameOver: true);
     }
 }